+++ src/include/benchmark.h
@@ -9,8 +9,8 @@ struct benchmark_params {
 	trx_mode modem;
 	int freq;
 	bool afc, sql;
-	double sqlevel;
-	double src_ratio;
+	float sqlevel;
+	float src_ratio;
 	int src_type;
 	std::string input, output, buffer;
 	size_t samples;
