+++ src/include/dominoex.h
@@ -65,11 +65,11 @@ public:
 	};
 protected:
 // common variables
-	double	phase[MAXFFTS + 1];
-	double	txphase;
+	float	phase[MAXFFTS + 1];
+	float	txphase;
 	int		symlen;
 	int		doublespaced;
-	double	tonespacing;
+	float	tonespacing;
 	int		counter;
 	unsigned int	twosym;
 	int		paths;
@@ -89,8 +89,8 @@ protected:
 	
 	domrxpipe		*pipe;
 	unsigned int	pipeptr;
-	mbuffer<double, 0, 2>	scopedata;
-	mbuffer<double, 0, 2>	videodata;
+	mbuffer<float, 0, 2>	scopedata;
+	mbuffer<float, 0, 2>	videodata;
 
 	cmplx currvector;
 
@@ -98,11 +98,11 @@ protected:
 	int prev1symbol;
 	int prev2symbol;
 
-	double met1;
-	double met2;
-	double sig;
-	double noise;
-	double s2n;
+	float met1;
+	float met2;
+	float sig;
+	float noise;
+	float s2n;
 
 	int synccounter;
 
@@ -168,7 +168,7 @@ public:
 	void	rx_init();
 	void	tx_init(SoundBase *sc);
 	void	restart();
-	int		rx_process(const double *buf, int len);
+	int		rx_process(const float *buf, int len);
 	int		tx_process();
 };
